refactor: notifier
